Move the TextUtils code into a much smaller TextUtils class in commons-db
