+++ commons-db/src/test/java/com/palantir/nexus/db/sql/TextUtilsTest.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright 2016 Palantir Technologies
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.palantir.nexus.db.sql;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class TextUtilsTest {
+
+    @Test
+    public void testRemoveAllWhitespace() {
+        String before = "  \r\n\n\r  \t FOOOooo\r\n\n\n\r\t\r o   \n";
+        String after = TextUtils.removeAllWhitespace(before);
+        assertEquals("FOOOoooo", after);
+    }
+}
